Move all test dependencies to separate module
