+++ .github/workflows/checks.yml
@@ -34,6 +34,22 @@ jobs:
     - name: Checkout code
       uses: actions/checkout@v2
 
+    - name: Run unit tests
+      run: go test -v -count=3
+
+  integration:
+    name: Integration Tests
+    runs-on: ubuntu-20.04
+
+    steps:
+    - name: Setup Go
+      uses: actions/setup-go@v2
+      with:
+        go-version: 1.16.x
+
+    - name: Checkout code
+      uses: actions/checkout@v2
+
     - name: Restore cache
       uses: actions/cache@v2
       with:
@@ -41,7 +57,7 @@ jobs:
         key: ${{ runner.os }}-build-${{ hashFiles('**/go.sum') }}
 
     - name: Build Message DB Docker image
-      run: docker build -t message-db .
+      run: docker build -t message-db -f tests/Dockerfile .
 
     - name: Run Message DB
       run: |
@@ -52,4 +68,4 @@ jobs:
           -c message_store.sql_condition=on
 
     - name: Run unit tests
-      run: go test -v -count=3 -condition-on
+      run: cd tests; go test -v -count=3 -condition-on
