--- client_test.go
@@ -1,777 +1,0 @@
-package gomdb_test
-
-import (
-	"context"
-	"database/sql"
-	"errors"
-	"flag"
-	"fmt"
-	"strconv"
-	"sync"
-	"testing"
-	"time"
-
-	"github.com/alexrudd/gomdb"
-	_ "github.com/lib/pq"
-	"github.com/thanhpk/randstr"
-)
-
-var (
-	host          = flag.String("host", "localhost", "the test db host")
-	port          = flag.Int("port", 5432, "the test db port")
-	dbname        = flag.String("dbname", "message_store", "the message-db database")
-	user          = flag.String("user", "message_store", "the user to connect as")
-	password      = flag.String("password", "", "the password to use to login")
-	sslmode       = flag.String("sslmode", "disable", "the ssl mode to connect with")
-	isConditionOn = flag.Bool("condition-on", false, "is the SQL condition feature on")
-)
-
-func Init() {
-	flag.Parse()
-}
-
-// NewClient opens a new DB connection then creates and returns a Client.
-func NewClient(t *testing.T) *gomdb.Client {
-	t.Helper()
-
-	conn := fmt.Sprintf("host=%s port=%v dbname=%s user=%s sslmode=%s",
-		*host, *port, *dbname, *user, *sslmode)
-
-	if *password != "" {
-		conn += " password=" + *password
-	}
-
-	db, err := sql.Open("postgres", conn)
-	if err != nil {
-		t.Fatalf("opening db (%s): %s", conn, err)
-	}
-
-	t.Cleanup(func() {
-		db.Close()
-	})
-
-	_, err = db.Exec("SET search_path TO message_store,public;")
-	if err != nil {
-		t.Fatalf("setting search path: %s", err)
-	}
-
-	return gomdb.NewClient(db)
-}
-
-// FakeUUID generates a fake but valid uuid for testing (this avoids importing
-// a uuid package).
-var monoclock = 0
-
-func FakeUUID() string {
-	monoclock++
-	return "abcd1234-ab12-cd34-" + fmt.Sprintf("%04d", monoclock) + "-" + strconv.FormatInt(time.Now().UnixNano(), 16)[4:]
-}
-
-// NewTestStream creates a new StreamIdentifier using the provided category
-// prefix.
-func NewTestStream(category string) gomdb.StreamIdentifier {
-	return gomdb.StreamIdentifier{
-		Category: category,
-		ID:       randstr.Base62(10),
-	}
-}
-
-// NewTestCategory returns a unique category name
-func NewTestCategory(prefix string) string {
-	return prefix + randstr.Base62(10)
-}
-
-// PopulateStream creates the specified number of messages and writes them
-// to the specified stream.
-func PopulateStream(t *testing.T, client *gomdb.Client, stream gomdb.StreamIdentifier, messages int) {
-	t.Helper()
-
-	var (
-		version = gomdb.NoStreamVersion
-		err     error
-	)
-
-	for i := 0; i < messages; i++ {
-		version, err = client.WriteMessage(context.TODO(), stream, gomdb.ProposedMessage{
-			ID:   FakeUUID(),
-			Type: "TestMessage",
-			Data: "data",
-		}, version)
-		if err != nil {
-			t.Fatal(err)
-		}
-	}
-}
-
-// PopulateCategory creates multiple streams within a single categatory and
-// populates them with messages. The actual category is returned.
-func PopulateCategory(t *testing.T, client *gomdb.Client, category string, streams, messages int) string {
-	t.Helper()
-
-	for i := 0; i < streams; i++ {
-		stream := gomdb.StreamIdentifier{
-			Category: category,
-			ID:       randstr.Base62(10),
-		}
-
-		PopulateStream(t, client, stream, messages)
-	}
-
-	return category
-}
-
-// TestWriteMessage tests the WriteMessage API.
-func TestWriteMessage(t *testing.T) {
-	t.Parallel()
-
-	client := NewClient(t)
-
-	t.Run("stream does not exist", func(t *testing.T) {
-		t.Parallel()
-
-		stream := NewTestStream(NewTestCategory("new_stream"))
-		msg := gomdb.ProposedMessage{
-			ID:   FakeUUID(),
-			Type: "TestMessage",
-			Data: "data",
-		}
-
-		version, err := client.WriteMessage(context.TODO(), stream, msg, gomdb.NoStreamVersion)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		if version != 0 {
-			t.Fatalf("expected ver 0, got version %v", version)
-		}
-	})
-
-	t.Run("skip OCC check", func(t *testing.T) {
-		t.Parallel()
-
-		stream := NewTestStream(NewTestCategory("any_stream"))
-		msg := gomdb.ProposedMessage{
-			ID:   FakeUUID(),
-			Type: "TestMessage",
-			Data: "data",
-		}
-
-		// write initial message
-		_, _ = client.WriteMessage(context.TODO(), stream, msg, gomdb.NoStreamVersion)
-
-		// write any version
-		msg.ID = FakeUUID()
-		version, err := client.WriteMessage(context.TODO(), stream, msg, gomdb.AnyVersion)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		if version != 1 {
-			t.Fatalf("expected ver 1, got version %v", version)
-		}
-	})
-
-	t.Run("fail OCC check", func(t *testing.T) {
-		t.Parallel()
-
-		stream := NewTestStream(NewTestCategory("any_stream"))
-		msg := gomdb.ProposedMessage{
-			ID:   FakeUUID(),
-			Type: "TestMessage",
-			Data: "data",
-		}
-
-		// write initial message
-		_, _ = client.WriteMessage(context.TODO(), stream, msg, gomdb.NoStreamVersion)
-
-		// write to same version version
-		msg.ID = FakeUUID()
-		_, err := client.WriteMessage(context.TODO(), stream, msg, gomdb.NoStreamVersion)
-		if !errors.Is(err, gomdb.ErrUnexpectedStreamVersion) {
-			t.Fatal("expected OCC failure")
-		}
-	})
-}
-
-// TestGetStreamMessages tests the GetStreamMessages API.
-func TestGetStreamMessages(t *testing.T) {
-	t.Parallel()
-
-	client := NewClient(t)
-
-	t.Run("stream does not exist", func(t *testing.T) {
-		t.Parallel()
-
-		stream := NewTestStream(NewTestCategory("nonexistant"))
-
-		msgs, err := client.GetStreamMessages(context.TODO(), stream)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		if len(msgs) != 0 {
-			t.Fatalf("expected no messages, got %v", len(msgs))
-		}
-	})
-
-	t.Run("get entire stream", func(t *testing.T) {
-		t.Parallel()
-
-		stream := NewTestStream(NewTestCategory("entire"))
-		PopulateStream(t, client, stream, 10)
-
-		msgs, err := client.GetStreamMessages(context.TODO(), stream)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		if len(msgs) != 10 {
-			t.Fatalf("expected 10 messages, got %v", len(msgs))
-		}
-	})
-
-	t.Run("get first half of stream", func(t *testing.T) {
-		t.Parallel()
-
-		stream := NewTestStream(NewTestCategory("half"))
-		PopulateStream(t, client, stream, 10)
-
-		msgs, err := client.GetStreamMessages(context.TODO(), stream, gomdb.WithStreamBatchSize(5))
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		if len(msgs) != 5 {
-			t.Fatalf("expected 5 messages, got %v", len(msgs))
-		}
-
-		for ver, msg := range msgs {
-			if msg.Version != int64(ver) {
-				t.Fatalf("expected first message with version %v, got %v", ver, msg.Version)
-			}
-		}
-	})
-
-	t.Run("get second half of stream", func(t *testing.T) {
-		t.Parallel()
-
-		stream := NewTestStream(NewTestCategory("half"))
-		PopulateStream(t, client, stream, 10)
-
-		msgs, err := client.GetStreamMessages(context.TODO(), stream, gomdb.WithStreamBatchSize(5), gomdb.FromVersion(5))
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		if len(msgs) != 5 {
-			t.Fatalf("expected 10 messages, got %v", len(msgs))
-		}
-
-		for ver, msg := range msgs {
-			if msg.Version != int64(ver+5) {
-				t.Fatalf("expected first message with version %v, got %v", ver+5, msg.Version)
-			}
-		}
-	})
-
-	t.Run("get stream with condition", func(t *testing.T) {
-		t.Parallel()
-
-		if !*isConditionOn {
-			t.Skip()
-		}
-
-		stream := NewTestStream(NewTestCategory("conditional"))
-		PopulateStream(t, client, stream, 10)
-
-		msgs, err := client.GetStreamMessages(context.TODO(), stream,
-			gomdb.WithStreamCondition("MOD(messages.position, 2) = 0"),
-		)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		if len(msgs) != 5 {
-			t.Fatalf("expected 10 messages, got %v", len(msgs))
-		}
-
-		for _, msg := range msgs {
-			if msg.Version%2 != 0 {
-				t.Fatalf("expected only even versions, got %v", msg.Version)
-			}
-		}
-	})
-}
-
-// TestGetCategoryMessages tests the GetCategoryMessages API.
-func TestGetCategoryMessages(t *testing.T) {
-	t.Parallel()
-
-	client := NewClient(t)
-
-	t.Run("stream does not exist", func(t *testing.T) {
-		t.Parallel()
-
-		msgs, err := client.GetCategoryMessages(context.TODO(), "nonexistant")
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		if len(msgs) != 0 {
-			t.Fatalf("expected no messages, got %v", len(msgs))
-		}
-	})
-
-	t.Run("get all messages for category", func(t *testing.T) {
-		t.Parallel()
-
-		category := PopulateCategory(t, client, NewTestCategory("category"), 5, 10)
-
-		msgs, err := client.GetCategoryMessages(context.TODO(), category)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		if len(msgs) != 50 {
-			t.Fatalf("expected 50 messages, got %v", len(msgs))
-		}
-	})
-
-	t.Run("get half messages for category", func(t *testing.T) {
-		t.Parallel()
-
-		category := PopulateCategory(t, client, NewTestCategory("half"), 5, 10)
-
-		// read all messages.
-		msgs, _ := client.GetCategoryMessages(context.TODO(), category)
-		middlePos := msgs[25].GlobalPosition
-
-		// re-read half messages and compare
-		halfMsgs, err := client.GetCategoryMessages(context.TODO(), category, gomdb.FromPosition(middlePos))
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		for i, msg := range halfMsgs {
-			if msg.GlobalPosition != msgs[25+i].GlobalPosition {
-				t.Fatalf("expected position %v messages, got %v", msgs[25+i].GlobalPosition, msg.GlobalPosition)
-			}
-		}
-	})
-
-	t.Run("read only first 15 messages in category", func(t *testing.T) {
-		t.Parallel()
-
-		category := PopulateCategory(t, client, NewTestCategory("batched"), 5, 10)
-
-		msgs, err := client.GetCategoryMessages(context.TODO(), category, gomdb.WithCategoryBatchSize(15))
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		if len(msgs) != 15 {
-			t.Fatalf("expected 15 messages, got %v", len(msgs))
-		}
-	})
-
-	t.Run("read as consumer group", func(t *testing.T) {
-		t.Parallel()
-
-		category := PopulateCategory(t, client, NewTestCategory("consumer"), 5, 10)
-
-		msgs1, err := client.GetCategoryMessages(context.TODO(), category, gomdb.AsConsumerGroup(0, 2))
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		msgs2, err := client.GetCategoryMessages(context.TODO(), category, gomdb.AsConsumerGroup(1, 2))
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		msgs := append(msgs1, msgs2...)
-
-		if len(msgs) != 50 {
-			t.Fatalf("expected 50 messages, got %v", len(msgs))
-		}
-	})
-
-	t.Run("read with correlation", func(t *testing.T) {
-		t.Parallel()
-
-		category := NewTestCategory("correlation")
-		stream := NewTestStream(category)
-
-		// write correlated event
-		_, _ = client.WriteMessage(context.TODO(), stream, gomdb.ProposedMessage{
-			ID:   FakeUUID(),
-			Type: "Correlated",
-			Data: "data",
-			Metadata: map[string]string{
-				gomdb.CorrelationKey: "correlated",
-			},
-		}, gomdb.AnyVersion)
-
-		// write uncorrelated event
-		_, _ = client.WriteMessage(context.TODO(), stream, gomdb.ProposedMessage{
-			ID:   FakeUUID(),
-			Type: "Uncorrelated",
-			Data: "data",
-			Metadata: map[string]string{
-				gomdb.CorrelationKey: "uncorrelated",
-			},
-		}, gomdb.AnyVersion)
-
-		msgs, err := client.GetCategoryMessages(context.TODO(), category, gomdb.WithCorrelation("correlated"))
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		if len(msgs) != 1 {
-			t.Fatalf("expected 1 correlated message, actual: %v", len(msgs))
-		}
-
-		if msgs[0].Type != "Correlated" {
-			t.Fatalf("expected message type of Correlated, actual: %s", msgs[0].Type)
-		}
-
-	})
-
-	t.Run("read with condition", func(t *testing.T) {
-		t.Parallel()
-
-		if !*isConditionOn {
-			t.Skip()
-		}
-
-		category := PopulateCategory(t, client, NewTestCategory("condition"), 5, 10)
-
-		msgs, err := client.GetCategoryMessages(context.TODO(), category,
-			gomdb.WithCategoryCondition("MOD(messages.position, 2) = 0"),
-		)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		if len(msgs) != 25 {
-			t.Fatalf("expected 10 messages, got %v", len(msgs))
-		}
-
-		for _, msg := range msgs {
-			if msg.Version%2 != 0 {
-				t.Fatalf("expected only even versions, got %v", msg.Version)
-			}
-		}
-	})
-}
-
-// TestGetLastStreamMessage tests the GetLastStreamMessage API.
-func TestGetLastStreamMessage(t *testing.T) {
-	t.Parallel()
-
-	client := NewClient(t)
-
-	t.Run("stream does not exist", func(t *testing.T) {
-		t.Parallel()
-
-		stream := NewTestStream(NewTestCategory("nonexistant"))
-
-		msg, err := client.GetLastStreamMessage(context.TODO(), stream)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		if msg != nil {
-			t.Fatalf("expected no message, got %v", msg)
-		}
-	})
-
-	t.Run("get last message", func(t *testing.T) {
-		t.Parallel()
-
-		stream := NewTestStream(NewTestCategory("stream"))
-		PopulateStream(t, client, stream, 3)
-
-		msg, err := client.GetLastStreamMessage(context.TODO(), stream)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		if msg == nil {
-			t.Fatal("expected message but got nil")
-		} else if msg.Version != 2 {
-			t.Fatalf("expected message with version %v, actual %v", 2, msg.Version)
-		}
-	})
-}
-
-// TestGetStreamVersion tests the GetStreamVersion API.
-func TestGetStreamVersion(t *testing.T) {
-	t.Parallel()
-
-	client := NewClient(t)
-
-	t.Run("stream does not exist", func(t *testing.T) {
-		t.Parallel()
-
-		stream := NewTestStream(NewTestCategory("nonexistant"))
-
-		version, err := client.GetStreamVersion(context.TODO(), stream)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		if version != -1 {
-			t.Fatalf("expected version -1, actual %v", version)
-		}
-	})
-
-	t.Run("get stream version", func(t *testing.T) {
-		t.Parallel()
-
-		stream := NewTestStream(NewTestCategory("stream"))
-		PopulateStream(t, client, stream, 3)
-
-		version, err := client.GetStreamVersion(context.TODO(), stream)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		if version != 2 {
-			t.Fatalf("expected stream version %v, actual %v", 2, version)
-		}
-	})
-}
-
-// TestSubscribeToStream tests the SubscribeToStream API.
-func TestSubscribeToStream(t *testing.T) {
-	t.Parallel()
-
-	client := NewClient(t)
-
-	t.Run("subscribe to empty stream", func(t *testing.T) {
-		t.Parallel()
-
-		stream := NewTestStream(NewTestCategory("empty"))
-		ctx, cancel := context.WithCancel(context.TODO())
-		defer cancel()
-
-		goneLive := sync.WaitGroup{}
-		goneLive.Add(1)
-
-		err := client.SubscribeToStream(
-			ctx,
-			stream,
-			func(m *gomdb.Message) {
-				t.Fatal("No messages should exist on stream")
-			},
-			func(live bool) {
-				if !live {
-					t.Fatal("subscription should be live")
-				}
-				goneLive.Done()
-			},
-			func(err error) {
-				if err != nil {
-					t.Fatalf("received subscription error: %s", err)
-				}
-			},
-		)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		goneLive.Wait()
-	})
-
-	t.Run("subscribe to new messages", func(t *testing.T) {
-		t.Parallel()
-
-		stream := NewTestStream(NewTestCategory("incoming"))
-		ctx, cancel := context.WithCancel(context.TODO())
-		defer cancel()
-
-		received := sync.WaitGroup{}
-		received.Add(3)
-
-		err := client.SubscribeToStream(
-			ctx,
-			stream,
-			func(m *gomdb.Message) {
-				received.Done()
-			},
-			func(live bool) {},
-			func(err error) {},
-		)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		PopulateStream(t, client, stream, 3)
-
-		received.Wait()
-	})
-
-	t.Run("catch up to stream then go live", func(t *testing.T) {
-		t.Parallel()
-
-		stream := NewTestStream(NewTestCategory("catchup"))
-		PopulateStream(t, client, stream, 10)
-
-		ctx, cancel := context.WithCancel(context.TODO())
-		defer cancel()
-
-		received := sync.WaitGroup{}
-		received.Add(10)
-		version := int64(0)
-
-		err := client.SubscribeToStream(
-			ctx,
-			stream,
-			func(m *gomdb.Message) {
-				version = m.Version
-				received.Done()
-			},
-			func(live bool) {
-				if live && version != 9 {
-					t.Fatalf("expected to go live at version 9, actual: %v", version)
-				}
-			},
-			func(err error) {},
-			gomdb.WithStreamBatchSize(5),
-		)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		received.Wait()
-	})
-}
-
-// TestSubscribeToCategory tests the SubscribeToCategory API.
-func TestSubscribeToCategory(t *testing.T) {
-	t.Parallel()
-
-	client := NewClient(t)
-
-	t.Run("subscribe to empty category", func(t *testing.T) {
-		t.Parallel()
-
-		ctx, cancel := context.WithCancel(context.TODO())
-		defer cancel()
-
-		goneLive := sync.WaitGroup{}
-		goneLive.Add(1)
-
-		err := client.SubscribeToCategory(
-			ctx,
-			NewTestCategory("empty"),
-			func(m *gomdb.Message) {
-				t.Fatal("No messages should exist on stream")
-			},
-			func(live bool) {
-				if !live {
-					t.Fatal("subscription should be live")
-				}
-				goneLive.Done()
-			},
-			func(err error) {
-				if err != nil {
-					t.Fatalf("received subscription error: %s", err)
-				}
-			},
-		)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		goneLive.Wait()
-	})
-
-	t.Run("subscribe to new category messages", func(t *testing.T) {
-		t.Parallel()
-
-		ctx, cancel := context.WithCancel(context.TODO())
-		defer cancel()
-
-		goneLive := sync.WaitGroup{}
-		goneLive.Add(1)
-		received := sync.WaitGroup{}
-		received.Add(30)
-
-		category := NewTestCategory("empty")
-
-		err := client.SubscribeToCategory(
-			ctx,
-			category,
-			func(m *gomdb.Message) {
-				received.Done()
-			},
-			func(live bool) {
-				if !live {
-					t.Fatal("subscription should be live")
-				}
-				goneLive.Done()
-			},
-			func(err error) {
-				if err != nil {
-					t.Fatalf("received subscription error: %s", err)
-				}
-			},
-		)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		goneLive.Wait()
-
-		PopulateCategory(t, client, category, 3, 10)
-
-		received.Wait()
-	})
-
-	t.Run("catch up to category then go live", func(t *testing.T) {
-		t.Parallel()
-
-		ctx, cancel := context.WithCancel(context.TODO())
-		defer cancel()
-
-		goneLive := sync.WaitGroup{}
-		goneLive.Add(1)
-		received := sync.WaitGroup{}
-		received.Add(30)
-
-		category := NewTestCategory("empty")
-		PopulateCategory(t, client, category, 3, 10)
-
-		err := client.SubscribeToCategory(
-			ctx,
-			category,
-			func(m *gomdb.Message) {
-				received.Done()
-			},
-			func(live bool) {
-				if !live {
-					t.Fatal("subscription should be live")
-				}
-				goneLive.Done()
-			},
-			func(err error) {
-				if err != nil {
-					t.Fatalf("received subscription error: %s", err)
-				}
-			},
-		)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		received.Wait()
-		goneLive.Wait()
-
-		// receive 10 more messages live
-		received.Add(10)
-		PopulateCategory(t, client, category, 2, 5)
-		received.Wait()
-	})
-}
