+++ go.mod
@@ -1,10 +1,3 @@
 module github.com/alexrudd/gomdb
 
 go 1.16
-
-require (
-	github.com/gofrs/uuid v1.2.0
-	github.com/lib/pq v1.10.1 // test dep
-	github.com/thanhpk/randstr v1.0.4 // test dep
-	gopkg.in/check.v1 v1.0.0-20201130134442-10cb98267c6c // indirect
-)
